Deleted deprecated backend apps and adjusted frontend (App.js, CourseMapping, Dashboard) for updated navigation
